fix: unit test error
